Added new api call to replace a half dozen
Previously, searching invloved calling a service to convert rss to json,
then calling 6 api endpoints for the rest of the information about the
posts. This was slow, and major source of failure due to the api server
being down, and the the rss -> json service going out of business. This
commit is just an api endpoint to get the info about the last x posts
from y city.