Complete overhaul of /search
In the intervening months since I worked on this project two things
happened: blastcasta went down, and I realized my results api was
mediorce.

In response, I added an api endpoint, see 55a0831, which returns all the
data about the most recent postings. This reduced the six requests to
one. This causes the page to load faster, and more reliably as the
weakest point of the salt/clrides has been the results api.

Search.erb was divided into three parts, haml, less, and coffee script.
The refactoring reduced the amount of requests, and includes more
idiomatic jQuery. Because of the separation of concerns, the banner
image is based on the class of the banner container. The less file
handles the matching between place and banner image. The banner images
were moved to /img/banner

While /search uses haml, layout is still in erb. basics.rb was changed
to support the differing rendering engines.