Changes to make firefox work better and search
-webkit-background-clip: text is the property that makes the names in
today.erb look so cool. Unfortunately, firefox does not support this.
The solution is to set the background to be transparent in the place of
the image by using @-moz-document url-prefix(){} to specifically
target the gecko rendering engine.

I also removed the search button on the search page. It jutted out
further than the edge of the page, making the horizonal scrollbar
appear in firefox. Furthermore, the search fields update the search
results, without the button.

The dropdown changes the origin search across all browsers.